+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Additions
 
 - Added JsonBuilder.add(key, Iterable<JsonModel>) to avoid having to call toJSON() on it (https://github.com/edvin/tornadofx/issues/414)
+- ViewModel partial rollback (https://github.com/edvin/tornadofx/issues/420)
 
 ## [1.7.10]